Windows Patches
src/assuan-handler.c:
(MSVC specific) MSVC can't construct an empty struct, therefore
default arguments to it.

src/w32-fd-t.inc.h:
Put GCC specific pragmas under ifdef __GNUC__

src/w32-types.inc.h:
Change ssize_t to 'long long'

+++ src/w32-types.inc.h
@@ -23,7 +23,7 @@
 typedef void *assuan_msghdr_t;
 
 #ifdef _MSC_VER
-  typedef long ssize_t;
+  typedef long long ssize_t;
   typedef int  pid_t;
 #endif
